requirements.txt
amadeus==5.3.1
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.20.0
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
djangorestframework==3.13.1